Overhaul the `env` module, add an example
