+++ Cargo.toml
@@ -7,9 +7,12 @@ license = "Apache-2.0 OR MIT"
 repository = "https://github.com/BlackHoleFox/appleargs"
 description = "Obtain the current process' apple arguments"
 keywords = ["apple", "darwin", "macos", "args", "applep"]
-categories = ["os::macos-apis"] 
+categories = ["os::macos-apis"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+empty-on-unsupported = []
 
 [dependencies]
